chore: cleanup logger
